+++ example/get_android_dependencies.ps1
@@ -1,5 +1,7 @@
 $AneVersion = "2.1.0"
-$PlayerServicesVersion = "15.0.1"
+$PlayerServicesVersion = "16.0.1"
+$MapsVersion = "16.1.0"
+$LocationVersion = "16.0.0"
 $SupportV4Version = "27.1.0"
 
 $currentDir = (Get-Item -Path ".\" -Verbose).FullName
@@ -10,5 +12,5 @@ Invoke-WebRequest -OutFile "$currentDir\android_dependencies\org.greenrobot.even
 Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.code.gson.gson-2.8.4.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/misc/com.google.code.gson.gson-2.8.4.ane?raw=true
 Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.android.support.support-v4-$SupportV4Version.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/support/com.android.support.support-v4-$SupportV4Version.ane?raw=true
 Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-base-$PlayerServicesVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-base-$PlayerServicesVersion.ane?raw=true
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-location-$PlayerServicesVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-location-$PlayerServicesVersion1.ane?raw=true
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-maps-$PlayerServicesVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-maps-$PlayerServicesVersion.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-location-$LocationVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-location-$LocationVersion.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-maps-$MapsVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-maps-$MapsVersion.ane?raw=true