+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 2.1.0
 - iOS: Refactor
-- iOS: Updated to FreSwift 3.1.0
+- AND: Updated to Android Play Services to 16.0.1
 
 ### 2.0.0
 - AND: Updated to FreKotlin 1.6.0